Fix mount ino in readdir.
When we reply to fuse, we need to pass the ino, not the entry_idx.
This but a bit silly. As the entry_idx is of by 2. We where returning a
ino of 0 for the root directory and so it was discarded and not displayed
in `ls`.
However, a `lookup` of the file was working so a diff between the mounted
archive and the source was working and did detect the issue.