Make our entry store sorted.
Our entry store was correctly sorted by parent_idx.
But inside a directory range, entry was not sorted by name.
With the new API in Jubako, we can ask him to sort it for us.
(With a bit more work on our side to be sure that parent_idx and
first_child_id are not invalid.

This also us to don't care in which order we add entries are they are
sorted at the end. So we simplify our process.

As entry are sorted, we can use the binary search in jubako.