Use temporary file to create packs.
With new api of jubako, we can specify in the pack_info different path
than the one used to create the file.

So we can use a temporary file to create the directory and content packs
and then move the files when created.

(In fact, this is mainly a commit to adapt to new api, but we use it
to add this improvement)