feat!: Better extract filter.
The previous version was a bit messy as user may wrongly set recurse value
depending of `-L` option or simple file given in command line.

And the code was a bit too complex as we had to accept parent directories
to be able to extract files.

Here we have a clear three options :
- From a file list (`-L`): No recursion, no globing... Only plain listing
- From command line argument: Recursion (if this is a directory), no globing
  Tell what you want
- From glob (`-g`, `--glob`): A glob pattern to select what to extract.