Use only one `HashMap<Children>` in dir entry at creation.
By merging `dir_children` and `file_children` into only one `children`,
we can catch any error as adding a link (to a directory) `foo` and
then adding a entry `foo/bar`.

Without this, when adding `foo/bar`, `foo` directory is created and,
we sorting the entry store at end of creation, we never succeed to sort
it as we have two `foo` entries (at least we fail to sort it and detect
the error).

Here we detect the error as soon has we try to add `foo/bar`.