feat!: Better api to create archive.
Instead of providing a strip_prefix, now arx trim all parents from given
path by default.
We add an option `--keep-parents` to use the full path. And `--dir-as-root`
to trim the directory itself (and so put files in the directory directly
in the root of the archive).

In the process, we now accept absolute path.